Prompts are now common config
